Completely rework ast.h and ast.c
Many new node types have been added, and each node type is now a struct
instead of having one giant tagged union. On the implementation side,
this means having a lot of macros to do most of the casting and other
repetitive work, but the result is not too difficult to work with.

Some allocations are made without checking the result; this is on
purpose, as in the near future I intend to introduce a macro to handle
the allocation failure case instead of checking the result every time.