Add a very limited C specification and grammar
These are based on the freely-available ANSI C Lex specification and
Yacc grammar, with some significant changes (mostly removals). At this
point, even some very basic parts of the language (e.g. 'return' and
string literals) are missing, but will be added in the near future.

The Lex specification and the Yacc grammar can be found here:
http://www.quut.com/c/ANSI-C-grammar-l-1999.html
http://www.quut.com/c/ANSI-C-grammar-y-1999.html

Many thanks to the authors for making them available and allowing for
others to modify their work; both the specification and grammar have
been incredibly helpful so far as I'm trying to figure things out.