Add basic DeclSpecs validation
The DeclSpec enum has been replaced with preprocessor constants, and two
bitmask constants have been added to simplify categorization.

At the moment, the validation is done in the 'decl_specs'. It is quite
limited, only checking whether the combination of specifiers is valid.

The validation is a bit large, and should probably be moved to a
function somewhere. Unfortunately, I have no idea where it should go.

To complicate things a bit, YYERROR can only be called from rule
actions. There are two obvious solutions:
- Check the result of the validation function in the action and call
  YYERROR if the validation failed
- Keep track of errors in ParserState instead, avoiding YYERROR entirely